requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
PyJWT>=2.8.0
cryptography>=41.0.0
redis>=5.0.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
